# traffic-safety-openapi/public/api/docs/openapi.json
openapi: 3.0.0
info:
  title: 'Traffic Safety Platform API'
  description: 'API documentation for the Traffic Safety decision management system.'
  version: 1.0.0
servers:
  -
    url: /
    description: 'Relative base URL'
paths:
...
  /api/countermeasures:
    get:
      tags:
        - Countermeasures
      summary: 'List countermeasures.'
      operationId: listCountermeasures
      responses:
        '200':
          description: 'Countermeasure collection.'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Countermeasure' } }
                type: object
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Countermeasures
      summary: 'Create a countermeasure.'
      operationId: createCountermeasure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Countermeasure'
      responses:
        '201':
          description: 'Countermeasure created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Countermeasure'
        '422':
          description: 'Invalid payload.'
      security:
        -
          bearerAuth: []
  '/api/countermeasures/{id}':
    get:
      tags:
        - Countermeasures
      summary: 'Get a countermeasure by id.'
      operationId: getCountermeasure
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Countermeasure details.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Countermeasure'
        '404':
          description: 'Countermeasure not found.'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Countermeasures
      summary: 'Update a countermeasure.'
      operationId: updateCountermeasure
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Countermeasure'
      responses:
        '200':
          description: 'Countermeasure updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Countermeasure'
        '404':
          description: 'Countermeasure not found.'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Countermeasures
      summary: 'Delete a countermeasure.'
      operationId: deleteCountermeasure
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Countermeasure deleted.'
      security:
        -
          bearerAuth: []
  /api/countermeasures/hotspot:
    post:
      tags:
        - Countermeasures
      summary: 'Find countermeasures suitable for a hotspot.'
      operationId: findCountermeasuresForHotspot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - targetType
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 'Matching countermeasures.'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Countermeasure' } }
                  count: { type: integer }
                type: object
        '422':
          description: 'Invalid filter payload.'
      security:
        -
          bearerAuth: []
  /api/countermeasures/recalculate-cmf:
    post:
      tags:
        - Countermeasures
      summary: 'Recalculate countermeasure CMF metrics using project and accident context.'
      operationId: recalculateCountermeasureCmf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - project
                - accident
              properties:
                project:
                  type: object
                  additionalProperties: true
                accident:
                  type: object
                  additionalProperties: true
              type: object
      responses:
        '200':
          description: 'CMF recalculation triggered.'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string }
                  projectId: { type: integer, format: int64 }
                  countermeasureId: { type: integer, format: int64 }
                type: object
        '422':
          description: 'Invalid payload.'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    MonetaryAmount:
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: USD
      type: object
...
    Countermeasure:
      required:
        - id
        - name
        - targetType
        - cmf
        - lifecycleStatus
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        targetType:
          type: string
        affectedCollisionTypes:
          type: array
          items:
            type: string
        affectedSeverities:
          type: array
          items:
            type: string
        cmf:
          type: number
          format: float
        lifecycleStatus:
          type: string
        implementationCost:
          $ref: '#/components/schemas/MonetaryAmount'
        expectedAnnualSavings:
          type: number
          format: float
          nullable: true
        evidence:
          type: string
          nullable: true
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...